pyproject.toml
[build-system]
requires = ["setuptools>=80.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hoophigher"
version = "0.1.0"
description = "Terminal higher-or-lower NBA game scaffold"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "nba_api>=1.10.2",
    "pydantic-settings>=2.13.1",
    "sqlmodel>=0.0.38",
    "textual>=8.2.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-textual-snapshot>=1.0.0",
]

[project.scripts]
hoophigher = "hoophigher.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
